Adds support for connecting and staying connected to Mumble
Known Issues:
- PIPO does not appear to receive all expected messages from the
server. It is unknown if either the socket queue is not being
processed correctly or the messages are not being sent from the server
in the first place.
- Scheduling may be sub-optimal: Writing to the socket is not
scheduled in the main tokio::select! block, so any writes (currently,
only the two initial Version and Authentication messages sent when
connecting and the Ping message sent every 10 seconds) will
effectively block the rest of the Mumble task.